Added documentation to newly created LobbyHub functions as well as javascript 'view' functions
