+++ modemas.client/src/LobbyJoinView.jsx
@@ -1,12 +1,28 @@
 import { useEffect, useState } from "react";
 
-export default function LobbyJoinView({ onCreateLobby, onJoinLobby }) {
+/**
+ * View for creating a new lobby or joining an existing one.
+ */
+export default function LobbyJoinView({ connection }) {
     const [playerName, setPlayerName] = useState("");
     const [lobbyId, setLobbyId] = useState("");
 
+    const createLobby = async () => {
+        if (connection) {
+            await connection.invoke("CreateLobby");
+        }
+    };
+
+    const joinLobby = async (lobbyId, playerName) => {
+        if (connection && lobbyId && playerName) {
+            await connection.invoke("JoinLobby", lobbyId, playerName);
+        }
+    };
+
+
     return (
         <div>
-            <button onClick={() => onCreateLobby()}>Create Lobby</button>
+            <button onClick={() => createLobby()}>Create Lobby</button>
             <input
                 type="text"
                 placeholder="Enter your name"
@@ -19,7 +35,7 @@ export default function LobbyJoinView({ onCreateLobby, onJoinLobby }) {
                 value={lobbyId}
                 onChange={(e) => setLobbyId(e.target.value)}
             />
-            <button onClick={() => onJoinLobby(lobbyId, playerName)}>Join Lobby</button>
+            <button onClick={() => joinLobby(lobbyId, playerName)}>Join Lobby</button>
         </div>
     );
 }
